fix: исправление ошибок

+++ ewm-service/src/main/java/ru/practicum/ewm/categories/service/CategoryServiceImpl.java
@@ -9,6 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.categories.dto.*;
 import ru.practicum.ewm.categories.model.Category;
 import ru.practicum.ewm.categories.repository.CategoryRepository;
+import ru.practicum.ewm.event.repository.EventRepository;
+import ru.practicum.ewm.exception.ExistingValidationException;
 import ru.practicum.ewm.exception.NotFoundException;
 
 import java.util.List;
@@ -20,6 +22,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
     private final CategoryMapper categoryMapper;
 
     @Override
@@ -46,6 +49,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     public void remove(Long id) {
         isExistsCategoryById(id);
+        if(eventRepository.existsByCategory_Id(id)) {
+            throw new ExistingValidationException("Категория не доступна для удаления");
+        }
         // Обратите внимание: с категорией не должно быть связано ни одного события.
         log.info("Delete CategoryId={}", id);
         categoryRepository.deleteById(id);
